Proposed handling of the $role variable
On the file SamlAuth/Auth/SamlAuth.php, an object of SamlUserProvider is created and 3 parameters are passed, while the constructor expects 4. I couldn't find a way to extract the information on roles at the point I was looking on. This proposed change would leave the call in SamlAuth/Auth/SamlAuth.php unaltered while adding a default value that Kanboard requires.